+++ packages/types/src/modular.ts
@@ -3,6 +3,7 @@
 
 import {INetworkCommonModule} from '@subql/types-core';
 import {Data} from 'ejs';
+import type {Interface} from 'ethers/lib/utils';
 import {SubqlCustomDatasource, SubqlDatasource, SubqlRuntimeDatasource} from './project';
 
 export interface EthereumNetworkModule
@@ -15,4 +16,5 @@ export interface EthereumNetworkModule
     renderTemplate: (templatePath: string, outputPath: string, templateData: Data) => Promise<void>
   ): Promise<void>;
   parseContractPath(path: string): {name: string; rawName: string; path: string[]};
+  getAbiInterface(projectPath: string, abiFileName: string): Interface;
 }