+++ packages/node/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Enable ts strict model 
 ## [4.7.0] - 2024-07-01
 ### Changed
 - Update with `@subql/node-core`, `@subql/common-substrate`