+++ packages/node/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.7.0] - 2024-07-03
 ### Changed
 - bump with `@subql/common-ethereum` and `@subql/types-ethereum`
 
@@ -505,7 +506,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Init release
 
-[Unreleased]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.6.0...HEAD
+[Unreleased]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.7.0...HEAD
+[4.7.0]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.6.0...node-ethereum/4.7.0
 [4.6.0]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.5.2...node-ethereum/4.6.0
 [4.5.2]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.5.1...node-ethereum/4.5.2
 [4.5.1]: https://github.com/subquery/subql-ethereum/compare/node-ethereum/4.5.0...node-ethereum/4.5.1